extract view with logo
