+++ Wire-iOS/Sources/UserInterface/Company Login/PreBackendSwitchViewController.swift
@@ -24,7 +24,7 @@ protocol PreBackendSwitchViewControllerDelegate {
     func preBackendSwitchViewControllerDidComplete(_ url: URL)
 }
 
-class PreBackendSwitchViewController: AuthenticationStepViewController {
+final class PreBackendSwitchViewController: AuthenticationStepViewController {
     
     var authenticationCoordinator: AuthenticationCoordinator?
     var backendURL: URL?
@@ -39,32 +39,32 @@ class PreBackendSwitchViewController: AuthenticationStepViewController {
     static let informationBackgroundBlue = UIColor(red: 220/255, green: 237/255, blue: 248/255, alpha: 1)
     
     // MARK: - UI Elements
+    let wireLogoInfoView = WireLogoInfoView()
+//    let contentView = UIView()
+//
+//    let headerView: UIView = {
+//        let view = UIView()
+//        view.backgroundColor = .black
+//        return view
+//    }()
+//
+//    let progressContainerView: UIView = {
+//        let view = UIView()
+//        view.layer.cornerRadius = 40
+//        view.backgroundColor = UIColor(red: 50/255, green: 54/255, blue: 57/255, alpha: 1)
+//        view.layer.shadowOffset = .zero
+//        view.layer.shadowRadius = 5
+//        view.layer.shadowOpacity = 0.29
+//        view.layer.shadowColor = UIColor.black.cgColor
+//        view.accessibilityIdentifier = "ProgressView"
+//        return view
+//    }()
     
-    let contentView = UIView()
-    
-    let headerView: UIView = {
-        let view = UIView()
-        view.backgroundColor = .black
-        return view
-    }()
-    
-    let progressContainerView: UIView = {
-        let view = UIView()
-        view.layer.cornerRadius = 40
-        view.backgroundColor = UIColor(red: 50/255, green: 54/255, blue: 57/255, alpha: 1)
-        view.layer.shadowOffset = .zero
-        view.layer.shadowRadius = 5
-        view.layer.shadowOpacity = 0.29
-        view.layer.shadowColor = UIColor.black.cgColor
-        view.accessibilityIdentifier = "ProgressView"
-        return view
-    }()
-    
-    let wireLogo: UIImageView = {
-        let logo = UIImageView(image: UIImage(named: "wire-logo-letter"))
-        logo.accessibilityIdentifier = "ProgressView.Logo"
-        return logo
-    }()
+//    let wireLogo: UIImageView = {
+//        let logo = UIImageView(image: UIImage(named: "wire-logo-letter"))
+//        logo.accessibilityIdentifier = "ProgressView.Logo"
+//        return logo
+//    }()
 
     let progressView: TimedCircularProgressView = {
         let progress = TimedCircularProgressView()
@@ -76,26 +76,26 @@ class PreBackendSwitchViewController: AuthenticationStepViewController {
         return progress
     }()
     
-    let titleLabel: UILabel = {
-        let label = UILabel()
-        label.font = FontSpec(.large, .semibold).font!
-        label.textAlignment = .center
-        label.text = "login.sso.backend_switch.title".localized
-        label.accessibilityValue = label.text
-        label.textColor = .black
-        return label
-    }()
+//    let titleLabel: UILabel = {
+//        let label = UILabel()
+//        label.font = FontSpec(.large, .semibold).font!
+//        label.textAlignment = .center
+//        label.text = "login.sso.backend_switch.title".localized
+//        label.accessibilityValue = label.text
+//        label.textColor = .black
+//        return label
+//    }()
     
-    let subtitleLabel: UILabel = {
-        let label = UILabel()
-        label.font = FontSpec(.normal, .regular).font!
-        label.textAlignment = .center
-        label.numberOfLines = 0
-        label.text = "login.sso.backend_switch.subtitle".localized
-        label.accessibilityValue = label.text
-        label.textColor = .black
-        return label
-    }()
+//    let subtitleLabel: UILabel = {
+//        let label = UILabel()
+//        label.font = FontSpec(.normal, .regular).font!
+//        label.textAlignment = .center
+//        label.numberOfLines = 0
+//        label.text = "login.sso.backend_switch.subtitle".localized
+//        label.accessibilityValue = label.text
+//        label.textColor = .black
+//        return label
+//    }()
     
     let informationLabel: UILabel = {
         let label = UILabel()
@@ -129,83 +129,84 @@ class PreBackendSwitchViewController: AuthenticationStepViewController {
     }
     
     private func configureSubviews() {
-        view.addSubview(headerView)
+        view.addSubview(wireLogoInfoView)
+//        view.addSubview(headerView)
  
-        contentView.addSubview(progressContainerView)
-        contentView.addSubview(titleLabel)
-        contentView.addSubview(subtitleLabel)
-        contentView.addSubview(informationLabel)
+//        contentView.addSubview(progressContainerView)
+//        contentView.addSubview(titleLabel)
+//        contentView.addSubview(subtitleLabel)
+        wireLogoInfoView.contentView.addSubview(informationLabel)
         
-        progressContainerView.addSubview(progressView)
-        progressContainerView.addSubview(wireLogo)
+        wireLogoInfoView.progressContainerView.addSubview(progressView)
+//        progressContainerView.addSubview(wireLogo)
 
-        view.addSubview(contentView)
+//        view.addSubview(contentView)
     }
     
     private func createConstraints() {
-        disableAutoresizingMaskTranslation(for: [
-            headerView,
-            contentView,
-            progressContainerView,
-            wireLogo,
+        [
+//            headerView,
+//            contentView,
+//            progressContainerView,
+//            wireLogo,
+            wireLogoInfoView,
             progressView,
-            titleLabel,
-            subtitleLabel,
+//            titleLabel,
+//            subtitleLabel,
             informationLabel,
-        ])
+        ].disableAutoresizingMaskTranslation()
         
         NSLayoutConstraint.activate([
+            wireLogoInfoView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
+            wireLogoInfoView.topAnchor.constraint(equalTo: view.topAnchor),
+            wireLogoInfoView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
+            wireLogoInfoView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
+
             // header view
-            headerView.heightAnchor.constraint(equalTo: view.heightAnchor, multiplier: 0.3, constant: 0),
-            headerView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            headerView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-            headerView.topAnchor.constraint(equalTo: view.topAnchor),
-            headerView.bottomAnchor.constraint(equalTo: contentView.topAnchor),
+//            headerView.heightAnchor.constraint(equalTo: view.heightAnchor, multiplier: 0.3, constant: 0),
+//            headerView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
+//            headerView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
+//            headerView.topAnchor.constraint(equalTo: view.topAnchor),
+//            headerView.bottomAnchor.constraint(equalTo: contentView.topAnchor),
 
             // content view
-            contentView.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 24),
-            contentView.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -24),
-            contentView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
+//            contentView.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 24),
+//            contentView.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -24),
+//            contentView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
             
             // progress container view
-            progressContainerView.centerYAnchor.constraint(equalTo: contentView.topAnchor),
-            progressContainerView.centerXAnchor.constraint(equalTo: contentView.centerXAnchor),
-            progressContainerView.widthAnchor.constraint(equalToConstant: 80),
-            progressContainerView.heightAnchor.constraint(equalToConstant: 80),
+//            progressContainerView.centerYAnchor.constraint(equalTo: contentView.topAnchor),
+//            progressContainerView.centerXAnchor.constraint(equalTo: contentView.centerXAnchor),
+//            progressContainerView.widthAnchor.constraint(equalToConstant: 80),
+//            progressContainerView.heightAnchor.constraint(equalToConstant: 80),
             
             // wire logo
-            wireLogo.centerYAnchor.constraint(equalTo: progressContainerView.centerYAnchor),
-            wireLogo.centerXAnchor.constraint(equalTo: progressContainerView.centerXAnchor),
+//            wireLogo.centerYAnchor.constraint(equalTo: progressContainerView.centerYAnchor),
+//            wireLogo.centerXAnchor.constraint(equalTo: progressContainerView.centerXAnchor),
             
             // progress view
-            progressView.centerXAnchor.constraint(equalTo: progressContainerView.centerXAnchor),
-            progressView.centerYAnchor.constraint(equalTo: progressContainerView.centerYAnchor),
-            progressView.widthAnchor.constraint(equalTo: progressContainerView.widthAnchor),
-            progressView.heightAnchor.constraint(equalTo: progressContainerView.heightAnchor),
+            progressView.centerXAnchor.constraint(equalTo: wireLogoInfoView.progressContainerView.centerXAnchor),
+            progressView.centerYAnchor.constraint(equalTo: wireLogoInfoView.progressContainerView.centerYAnchor),
+            progressView.widthAnchor.constraint(equalTo: wireLogoInfoView.progressContainerView.widthAnchor),
+            progressView.heightAnchor.constraint(equalTo: wireLogoInfoView.progressContainerView.heightAnchor),
 
             // title label
-            titleLabel.topAnchor.constraint(equalTo: progressContainerView.bottomAnchor, constant: 16),
-            titleLabel.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
-            titleLabel.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
+//            titleLabel.topAnchor.constraint(equalTo: progressContainerView.bottomAnchor, constant: 16),
+//            titleLabel.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
+//            titleLabel.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
             
             // subtitle label
-            subtitleLabel.topAnchor.constraint(equalTo: titleLabel.bottomAnchor, constant: 14),
-            subtitleLabel.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
-            subtitleLabel.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
+//            subtitleLabel.topAnchor.constraint(equalTo: titleLabel.bottomAnchor, constant: 14),
+//            subtitleLabel.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
+//            subtitleLabel.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
             
             // information label
-            informationLabel.topAnchor.constraint(equalTo: subtitleLabel.bottomAnchor, constant: 10),
-            informationLabel.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
-            informationLabel.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
+            informationLabel.topAnchor.constraint(equalTo: wireLogoInfoView.subtitleLabel.bottomAnchor, constant: 10),
+            informationLabel.leadingAnchor.constraint(equalTo: wireLogoInfoView.contentView.leadingAnchor),
+            informationLabel.trailingAnchor.constraint(equalTo: wireLogoInfoView.contentView.trailingAnchor),
         ])
     }
 
-    private func disableAutoresizingMaskTranslation(for views: [UIView]) {
-        for view in views {
-            view.translatesAutoresizingMaskIntoConstraints = false
-        }
-    }
-    
     func executeErrorFeedbackAction(_ feedbackAction: AuthenticationErrorFeedbackAction) {
         // NO OP
     }
